[template] Cal cleanup
