+++ templates/cal/functions/confirm.ts
@@ -1,14 +1,14 @@
 'use server'
 import type { BuildFrameData, FrameActionPayload } from '@/lib/farcaster'
+import { loadGoogleFontAllVariants } from '@/sdk/fonts'
+import { FrameError } from '@/sdk/handlers'
 import type { Config, State } from '..'
+import { bookCall } from '../utils/cal'
+import { getEventId } from '../utils/cal'
+import { getCurrentAndFutureDate } from '../utils/date'
+import { extractDatesAndSlots } from '../utils/date'
 import PageView from '../views/AfterConfirm'
 import FailView from '../views/Failed'
-
-import { loadGoogleFontAllVariants } from '@/sdk/fonts'
-import { bookCall } from '../utils/bookCall'
-import { extractDatesAndSlots } from '../utils/extractDatesAndSlots'
-import { getCurrentAndFutureDate } from '../utils/getDays'
-import { getEventSlug } from '../utils/getEventSlug'
 
 export default async function confirm(
     body: FrameActionPayload,
@@ -25,7 +25,11 @@ export default async function confirm(
     }
 
     const buttonIndex = body.untrustedData.buttonIndex
+
     if (buttonIndex === 1) {
+        //! this logic, and the `errors` function seems like they shouldn't exist anymore
+        //! we removed NotSatisfied and converted it to FrameError
+        //! either this or the naming is off
         return {
             buttons: [
                 {
@@ -37,6 +41,7 @@ export default async function confirm(
             fonts: fonts,
         }
     }
+
     const dates = getCurrentAndFutureDate(30)
     const url = `https://cal.com/api/trpc/public/slots.getSchedule?input=${encodeURIComponent(
         JSON.stringify({
@@ -67,14 +72,19 @@ export default async function confirm(
     const date = datesArray[params.date]
 
     const email = body.untrustedData.inputText
-    const eventTypeId = await getEventSlug(config.username, params.duration)
-    await bookCall(
-        email?.split('@')[0] || '',
-        email!,
-        slotsResponse.result.data.json.slots[date][params.slot].time,
-        eventTypeId,
-        config.username
-    )
+    const eventTypeId = await getEventId(config.username, params.duration)
+
+    try {
+        await bookCall(
+            email?.split('@')[0] || '',
+            email!,
+            slotsResponse.result.data.json.slots[date][params.slot].time,
+            eventTypeId,
+            config.username
+        )
+    } catch (error) {
+        throw new FrameError('Error booking event.')
+    }
 
     return {
         buttons: [
