+++ templates/cal/functions/date.ts
@@ -1,13 +1,12 @@
 'use server'
 import type { BuildFrameData, FrameActionPayload } from '@/lib/farcaster'
+import { loadGoogleFontAllVariants } from '@/sdk/fonts'
+import { FrameError } from '@/sdk/handlers'
 import type { Config, State } from '..'
+import { getCurrentAndFutureDate } from '../utils/date'
+import { extractDatesAndSlots } from '../utils/date'
 import PageView from '../views/Date'
 import NextView from '../views/Slot'
-
-import { loadGoogleFontAllVariants } from '@/sdk/fonts'
-import { extractDatesAndSlots } from '../utils/extractDatesAndSlots'
-import { getCurrentAndFutureDate } from '../utils/getDays'
-import { FrameError } from '@/sdk/handlers'
 
 export default async function dateHanlder(
     body: FrameActionPayload,
