--- templates/cal/utils/dayjs.ts
@@ -1,8 +1,0 @@
-import dayjs from 'dayjs'
-import duration from 'dayjs/plugin/duration'
-import utc from 'dayjs/plugin/utc'
-
-dayjs.extend(duration)
-dayjs.extend(utc)
-
-export default dayjs
