+++ templates/cal/utils/nft.ts
@@ -1,6 +1,6 @@
-import { client, baseClient, opClient } from './client'
+//! allow for more evm chains
+import { baseClient, client, opClient } from './viem'
 
-// ABI for the balanceOf function of ERC-20 tokens
 const balanceOfABI721 = [
     {
         'constant': true,
@@ -22,52 +22,6 @@ const balanceOfABI721 = [
         'type': 'function',
     },
 ] as const
-
-export const balances721 = async (ownerAddresses: any, contract: any, chain: any) => {
-    // biome-ignore lint/style/useConst: <explanation>
-    let calls = []
-    let hasNFT = false
-
-    // biome-ignore lint/correctness/noUnusedVariables: <explanation>
-    let selectedClient
-    switch (chain) {
-        case 'ETH': {
-            selectedClient = client
-            break
-        }
-        case 'BASE': {
-            selectedClient = baseClient
-            break
-        }
-        case 'OP': {
-            selectedClient = opClient
-            break
-        }
-        default:
-            throw new Error('Unsupported chain')
-    }
-
-    // biome-ignore lint/style/useForOf: <explanation>
-    for (let i = 0; i < ownerAddresses.length; i++) {
-        calls.push({
-            address: contract,
-            abi: balanceOfABI721,
-            functionName: 'balanceOf',
-            args: [ownerAddresses[i]],
-        })
-    }
-    const results = await selectedClient.multicall({
-        contracts: calls,
-    })
-    // biome-ignore lint/style/useForOf: <explanation>
-    for (let i = 0; i < results.length; i++) {
-        if (results[i].result && results[i].result?.toString() !== '0') {
-            hasNFT = true
-            break
-        }
-    }
-    return hasNFT
-}
 
 const balanceOfABI1155 = [
     {
@@ -95,7 +49,64 @@ const balanceOfABI1155 = [
     },
 ] as const
 
-export const balancesERC1155 = async (
+const nameABI = [
+    {
+        constant: true,
+        inputs: [],
+        name: 'name',
+        outputs: [{ name: '', type: 'string' }],
+        type: 'function',
+    },
+]
+
+export const holdsErc721 = async (ownerAddresses: any, contract: any, chain: any) => {
+    const calls = []
+    let hasNFT = false
+
+    let selectedClient
+    switch (chain) {
+        case 'ETH': {
+            selectedClient = client
+            break
+        }
+        case 'BASE': {
+            selectedClient = baseClient
+            break
+        }
+        case 'OP': {
+            selectedClient = opClient
+            break
+        }
+        default:
+            throw new Error('Unsupported chain')
+    }
+
+    // biome-ignore lint/style/useForOf: <explanation>
+    for (let i = 0; i < ownerAddresses.length; i++) {
+        calls.push({
+            address: contract,
+            abi: balanceOfABI721,
+            functionName: 'balanceOf',
+            args: [ownerAddresses[i]],
+        })
+    }
+
+    const results = await selectedClient.multicall({
+        contracts: calls,
+    })
+
+    // biome-ignore lint/style/useForOf: <explanation>
+    for (let i = 0; i < results.length; i++) {
+        if (results[i].result && results[i].result?.toString() !== '0') {
+            hasNFT = true
+            break
+        }
+    }
+
+    return hasNFT
+}
+
+export const holdsErc1155 = async (
     ownerAddresses: any,
     contract: any,
     tokenId: any,
@@ -147,4 +158,38 @@ export const balancesERC1155 = async (
     }
 
     return hasNFT
+}
+
+export async function getName(contractAddress: any, chain: any): Promise<string> {
+    let selectedClient
+
+    switch (chain) {
+        case 'ETH': {
+            selectedClient = client
+            break
+        }
+        case 'BASE': {
+            selectedClient = baseClient
+            break
+        }
+        case 'OP': {
+            selectedClient = opClient
+            break
+        }
+        default:
+            throw new Error('Unsupported chain')
+    }
+    try {
+        let name: any = await selectedClient.readContract({
+            address: contractAddress,
+            abi: nameABI,
+            functionName: 'name',
+            args: [],
+        })
+        name = name.toString()
+        return name
+    } catch (error) {
+        console.error('Error getting contract name:', error)
+        return ''
+    }
 }
